Update yolo.py
I've noticed that the keypoint validation error the model reports is orders of magnitude different from the training error, and doesn't really change over time.

I believe this is due to a pass-by-reference bug, which changes the predictions made in eval mode:
When the model is set to eval mode, the IKeypoint.forward method runs through some extra steps. We reassign a slice of the original output to a new variable y, but since it's passed by reference the subsequent code changes the original outputs. Fixing this by using .clone.detach().